+++ backend/src/baserow/contrib/database/fields/registries.py
@@ -569,10 +569,11 @@ class FieldType(
             exporting the table. This is for example used by the link row field type
             to prefetch all relations.
         :type cache: dict
-        :param files_zip:
-        :type files_zip:
-        :param storage:
-        :type storage:
+        :param files_zip: A zip file buffer where the files related to the template
+            must be copied into.
+        :type files_zip: ZipFile
+        :param storage: The storage where the files can be loaded from.
+        :type storage: Storage or None
         :return: The exported value.
         :rtype: Object
         """
@@ -591,10 +592,11 @@ class FieldType(
         :type field_name: str
         :param value: The value that must be set.
         :type value: Object
-        :param files_zip:
-        :type files_zip:
-        :param storage:
-        :type storage:
+        :param files_zip: A zip file buffer where files related to the template can
+            be extracted from.
+        :type files_zip: ZipFile
+        :param storage: The storage where the files can be copied to.
+        :type storage: Storage or None
         :param id_mapping: The map of exported ids to newly created ids that must be
             updated when a new instance has been created.
         :type id_mapping: dict